refactor: change controller according to schemas

+++ src/app/controllers/channel.controller.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { NextFunction, Request, Response } from 'express';
-import { Channel } from '../schemas';
+import { Channel, Workspace } from '../schemas';
 import logger from '../util/logger';
 import { Types } from 'mongoose';
 
@@ -10,7 +10,7 @@ export async function listChannel(
   next: NextFunction,
 ) {
   try {
-    const channels = await Channel.find();
+    const channels = await Channel.find().populate('workspace');
     res.json(channels);
   } catch (error: any) {
     logger.error(error.message);
@@ -31,12 +31,31 @@ export async function createChannel(
       members.push(...req.body.members);
     }
 
-    const channel = await Channel.create({
-      name: req.body.name,
-      description: req.body.description,
-      owner: owner,
-      members: members,
+    const workspace = await Workspace.findOne({
+      name: req.body.workspace,
     });
+    if (!workspace) {
+      return res.status(404);
+    }
+
+    const channel = await Channel.findOneAndUpdate(
+      {
+        name: req.body.name,
+        workspace: workspace,
+      },
+      {
+        name: req.body.name,
+        description: req.body.description,
+        owner: owner,
+        members: members,
+        workspace: workspace,
+      },
+      {
+        upsert: true,
+        new: true,
+      },
+    );
+
     res.json(channel);
   } catch (error: any) {
     logger.error(error.message);
@@ -50,15 +69,17 @@ export async function deleteChannel(
   next: NextFunction,
 ) {
   try {
+    const workspace = await Workspace.findOne({
+      name: req.body.workspace,
+    });
+    if (!workspace) {
+      return res.status(404);
+    }
+
     const channel = await Channel.deleteOne({
-      $and: [
-        {
-          name: req.body.name,
-        },
-        {
-          owner: res.locals.user.id,
-        },
-      ],
+      name: req.body.name,
+      owner: res.locals.user.id,
+      workspace: workspace,
     });
     res.json(channel);
   } catch (error: any) {
@@ -77,10 +98,18 @@ export async function updateChannels(
     const addedMembers = req.body.addedMembers;
     const kickedMembers = req.body.kickedMembers;
 
+    const workspace = await Workspace.findOne({
+      name: req.body.workspace,
+    });
+    if (!workspace) {
+      return res.status(404);
+    }
+
     let channel = await Channel.findOneAndUpdate(
       {
         name: req.body.name,
         owner: owner,
+        workspace: workspace,
       },
       {
         description: req.body.description,
@@ -97,6 +126,7 @@ export async function updateChannels(
       {
         name: req.body.name,
         owner: owner,
+        workspace: workspace,
       },
       {
         $pull: {
